feat: Rename technicians api route (DSD-62)
