+++ src/app/api/technicians/by-department/[id]/route.ts
@@ -1,0 +1,30 @@
+import { createClient } from "@/utils/supabase/server";
+import { type NextRequest, NextResponse } from "next/server";
+
+export async function GET(
+  request: NextRequest,
+  { params }: { params: { id: string } },
+) {
+  const { id } = await params;
+  const departmentId = parseInt(id, 10);
+
+  if (isNaN(departmentId)) {
+    return NextResponse.json(
+      { error: "Invalid department ID" },
+      { status: 400 },
+    );
+  }
+
+  const supabase = await createClient();
+  const { data, error } = await supabase
+    .from("profiles")
+    .select("first_name, last_name, technician_details!inner(department_id)")
+    .eq("role", "TECHNICIAN")
+    .eq("technician_details.department_id", departmentId);
+
+  if (error) {
+    return NextResponse.json({ error: error.message }, { status: 500 });
+  }
+
+  return NextResponse.json(data);
+}
